call safe tDiv method with int argument

+++ src/de/tilman_neumann/jml/factor/cfrac/tdiv/TDiv_CF63_02.java
@@ -171,7 +171,7 @@ public class TDiv_CF63_02 implements TDiv_CF63 {
 		long factor1;
 		int Q_rest_bits = bitLength(Q_rest);
 		if (Q_rest_bits <= 24) {
-			factor1 = tDiv31.findSingleFactor(Q_rest);
+			factor1 = tDiv31.findSingleFactor((int) Q_rest);
 		} else if (Q_rest_bits <= 56) {
 			factor1 = lehman.findSingleFactor(Q_rest);
 		} else { // max Q_rest_bits is 63
